@@ -23,144 +23,213 @@
 # SOFTWARE.
 
 """
-Bases classes for concept explainers
+Bases Classes for Concept-based Explainers
 """
 
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Literal
+from collections.abc import Callable
+from functools import wraps
+from textwrap import dedent
+from typing import Any, Generic, TypeVar
 
 import torch
-from beartype import beartype
-from beartype.typing import List
-from jaxtyping import Float, Int, jaxtyped
+from overcomplete.base import BaseDictionaryLearning
+from torch import nn
 
 from interpreto.attributions.base import AttributionExplainer
-from interpreto.commons.model_wrapping.model_splitter import ModelSplitterPlaceholder
+from interpreto.commons.model_wrapping.model_with_split_points import ModelWithSplitPoints
 from interpreto.typing import ConceptsActivations, LatentActivations, ModelInput
 
+Module = TypeVar("Module", bound=nn.Module)
+T = TypeVar("T")
 
-class AbstractConceptExplainer(ABC):
-    """
-    Abstract class defining an interface for concept explanation.
+
+# Decorator that checks if the concept model is fitted before calling the method
+def check_fitted(func: Callable[..., T]) -> Callable[..., T]:
+    @wraps(func)
+    def wrapper(self: AbstractConceptExplainer, *args, **kwargs) -> T:
+        if not self.is_fitted or self.split_point is None:
+            raise RuntimeError("Concept encoder is not fitted yet. Use the .fit() method to fit the explainer.")
+        return func(self, *args, **kwargs)
+
+    return wrapper
+
+
+class AbstractConceptExplainer(ABC, Generic[Module]):
+    """Abstract class defining an interface for concept explanation.
+    Child classes should implement the `fit` and `encode_activations` methods, and only assume the presence of an
+        encoding step using the `concept_model` to convert activations to latent concepts.
 
     Attributes:
-        splitted_model (ModelSplitterPlaceholder): The model to apply the explanation on. It should be splitted between at least two parts.
-        split (str): The split in the model where the concepts are encoded from.
-        fitted (bool): Whether the model has been fitted
-        _differentiable_concept_encoder (bool): Whether the concept encoder is differentiable.
+        model_with_split_points (ModelWithSplitPoints): The model to apply the explanation on.
+            It should have at least one split point on which `concept_model` can be fitted.
+        split_point (str): The split point used to train the `concept_model`.
+        concept_model (torch.nn.Module): The model used to extract concepts from the activations of
+            `model_with_split_points`. The only assumption for classes inheriting from this class is that
+            the `concept_model` can encode activations into concepts with `encode_activations`.
+        is_fitted (bool): Whether the `concept_model` was fit on model activations.
+        has_differentiable_concept_encoder (bool): Whether the `encode_activations` operation is differentiable.
     """
 
-    _differentiable_concept_encoder: bool = False
-
-    def __init__(self, splitted_model: ModelSplitterPlaceholder):
-        """
-        Initializes the concept explainer with a given model.
-
-        Args:
-            splitted_model (ModelSplitterPlaceholder): The model to apply the explanation on. It should be splitted between at least two parts.
-        """
-        if not isinstance(splitted_model, ModelSplitterPlaceholder):
-            raise TypeError("Model should be a ModelSplitterPlaceholder.")
-        self.splitted_model: ModelSplitterPlaceholder = splitted_model
-
-    def verify_activations(
-        self, activations: dict[str, LatentActivations] | LatentActivations, split: str | None = None
-    ) -> tuple(LatentActivations, str | None):
-        """
-        Verify that the given activations are valid for the concept explainer.
-        That is if the split corresponds to the model splits.
-
-        Args:
-            activations (dict[str, LatentActivations]): the activations to verify.
-            split: (str | None): The dataset split to use for training the concept encoder. If None, the model is assumed to be a single-split model. And split is inferred from the keys of the activations dict.
-        """
-        if hasattr(self, "split") and self.split is not None:
-            split: str = self.split
-
-        if isinstance(activations, dict):
-            if len(activations) != len(self.splitted_model.splits):
-                raise ValueError(
-                    f"Activations should be a dict with {self.splitted_model.splits} keys but got {activations.keys()}."
-                )
-            if split is None:
-                if len(activations) != 1 and len(self.splitted_model.splits) != 1:
-                    raise ValueError("Cannot infer split if the model is a not a single-split model.")
-                split: str = list(activations.keys())[0]
-        elif split is None:
-            if len(self.splitted_model.splits) != 1:
-                raise ValueError("Cannot infer split if the model is a not a single-split model.")
-            split: str = self.splitted_model.splits[0]
-            # assuming the users knows what they are doing
-            return activations, None
-
-        assert split in activations, f"Split {split} not found in activations."
-        assert split in self.splitted_model.splits, f"Split {split} not found in model."
-
-        return activations[split], split
-
-    @abstractmethod
-    def fit(self, activations: dict[str, LatentActivations], split: str | None = None):
-        """
-        Defines the concept encoder, thus the concept space, using the given activations.
-
-        Args:
-            activations (dict[str, LatentActivations]): the activations to train the concept encoder on.
-            split: (str | None): The dataset split to use for training the concept encoder. If None, the model is assumed to be a single-split model. And split is inferred from the keys of the activations dict.
-
-        Returns:
-            The concept encoder.
-        """
-        pass
-
-    @abstractmethod
-    def encode_activations(
-        self, activations: LatentActivations | dict[str, LatentActivations], **kwargs
-    ) -> ConceptsActivations:
-        """
-        Encode the given activations using the concept encoder-decoder.
-
-        Args:
-            activations (LatentActivations | dict[str, LatentActivations]): The activations to encode.
-
-        Returns:
-            ConceptsActivations: The encoded activations.
-        """
-        pass
-
-    def top_k_inputs_for_concept(
-        self, inputs: ModelInput, concept: int | list[int], k: int = 5
-    ) -> list[tuple[str, float]]:
-        """
-        Retrieves the top-k most important tokens/words/clauses/phrases related to a given concept.
-
-        Args:
-            inputs: The input data, which can be a string, a list of tokens/words/clauses/sentences, or a dataset.
-            concept: The concept index (or list of concepts indices) to analyze.
-            k: The number of important textual elements to retrieve. Defaults to 5.
-
-        Returns:
-            A list of tuples containing the top-k most relevant textual elements and their importance scores.
-        """
-        print(inputs, concept, k)
-        raise NotImplementedError()
-
-    @jaxtyped
-    @beartype
-    def top_k_tokens_for_concept_from_activations(
+    def __init__(
         self,
-        inputs: List[List[str]],  # ModelInput
-        activations: Float[torch.Tensor, "n l d"],  # LatentActivations
+        model_with_split_points: ModelWithSplitPoints,
+        concept_model: Module,
+        split_point: str | None = None,
+    ):
+        """Initializes the concept explainer with a given splitted model.
+
+        Args:
+            model_with_split_points (ModelWithSplitPoints): The model to apply the explanation on.
+                It should have at least one split point on which a concept explainer can be trained.
+            concept_model (torch.nn.Module): The model used to extract concepts from
+                the activations of `model_with_split_points`.
+            split_point (str | None): The split point used to train the `concept_model`. If None, tries to use the
+                split point of `model_with_split_points` if a single one is defined.
+        """
+        if not isinstance(model_with_split_points, ModelWithSplitPoints):
+            raise TypeError(
+                f"The given model should be a ModelWithSplitPoints, but {type(model_with_split_points)} was given."
+            )
+        self.model_with_split_points: ModelWithSplitPoints = model_with_split_points
+        self.concept_model: Module = concept_model
+        self.split_point: str = self.get_and_verify_split_point(split_point)
+        self.is_fitted: bool = False
+        self.has_differentiable_concept_encoder = False
+
+    def __repr__(self):
+        return dedent(f"""\
+            {self.__class__.__name__}(
+                split_point={self.split_point},
+                concept_model={type(self.concept_model).__name__},
+                is_fitted={self.is_fitted},
+                has_differentiable_concept_encoder={self.has_differentiable_concept_encoder},
+            )""")
+
+    @abstractmethod
+    def fit(self, activations: LatentActivations | dict[str, LatentActivations], *args, **kwargs) -> Any:
+        """Fits `concept_model` on the given activations.
+
+        Args:
+            activations (dict[str, torch.Tensor]): A dictionary with model paths as keys and the corresponding
+                tensors as values.
+
+        Returns:
+            `None`, `concept_model` is fitted in-place, `is_fitted` is set to `True` and `split_point` is set.
+        """
+        pass
+
+    @abstractmethod
+    def encode_activations(self, activations: LatentActivations) -> ConceptsActivations:
+        """Abstract method defining how activations are converted into concepts by the concept encoder.
+
+        Args:
+            activations (torch.Tensor): The activations to encode.
+
+        Returns:
+            A `torch.Tensor` of encoded activations produced by the fitted concept encoder.
+        """
+        pass
+
+    def verify_activations(self, activations: dict[str, LatentActivations]) -> None:
+        """
+        Verify that the given activations are valid for the `model_with_split_points` and `self.split_point`.
+        Cases in which the activations are not valid include:
+
+        * Activations are not a valid dictionary.
+        * Specified split point does not exist in the activations.
+
+        Args:
+            activations (dict[str, torch.Tensor]): A dictionary with model paths as keys and the corresponding
+                tensors as values.
+        """
+        if not isinstance(activations, dict) or not all(isinstance(act, torch.Tensor) for act in activations.values()):
+            raise TypeError(
+                "Invalid activations for the concept explainer. "
+                "Activations should be a dictionary of model paths and torch.Tensor activations. "
+                f"Got: '{type(activations)}'"
+            )
+        activations_split_points = list(activations.keys())
+        if self.split_point not in activations_split_points:
+            raise ValueError(
+                f"Fitted split point '{self.split_point}' not found in activations.\n"
+                f"Available split_points: {', '.join(activations_split_points)}."
+            )
+
+    def get_and_verify_split_point(self, split_point: str | None) -> str:
+        if split_point is None and len(self.model_with_split_points.split_points) > 1:
+            raise ValueError(
+                "If the model has more than one split point, a split point for fitting the concept model should "
+                f"be specified. Got split point: '{split_point}' with model split points: "
+                f"{', '.join(self.model_with_split_points.split_points)}."
+            )
+        if split_point is None:
+            split_point = self.model_with_split_points.split_points[0]
+        if split_point not in self.model_with_split_points.split_points:
+            raise ValueError(
+                f"Split point '{split_point}' not found in model split points: "
+                f"{', '.join(self.model_with_split_points.split_points)}."
+            )
+        return split_point
+
+    def prepare_fit(
+        self,
+        activations: LatentActivations | dict[str, LatentActivations],
+        overwrite: bool,
+    ) -> LatentActivations:
+        if self.is_fitted and not overwrite:
+            raise RuntimeError(
+                "Concept explainer has already been fitted. Refitting will overwrite the current model."
+                "If this is intended, use `overwrite=True` in fit(...)."
+            )
+        if isinstance(activations, dict):
+            self.verify_activations(activations)
+            split_activations = activations[self.split_point]
+        else:
+            split_activations = activations
+        assert len(split_activations.shape) == 2, (
+            f"Input activations should be a 2D tensor of shape (batch_size, n_features) but got {split_activations.shape}."
+        )
+        return split_activations
+
+    @check_fitted
+    def top_k_inputs_for_concept(self, inputs: ModelInput, concept_id: int, k: int = 5) -> list[tuple[str, float]]:
+        """
+        Retrieves the top-k most important input elements (tokens/words/clauses/phrases) related to a given concept.
+
+        Args:
+            inputs (ModelInput): The input data, which can be a string, a list of tokens/words/clauses/sentences
+                or a dataset.
+            concept_id (int): Index identifying the position of the concept of interest (score in the
+                `ConceptActivations` tensor) for which relevant elements should be retrieved.
+            k: The number of input elements to retrieve. Defaults to 5.
+
+        Returns:
+            An ordered list of k tuples, each containing containing one of the top-k relevant input elements
+                for the specified concept and its importance scores.
+        """
+        self.split_point: str  # Verified by check_fitted
+        activations = self.model_with_split_points.get_activations(inputs)
+        return self.top_k_inputs_for_concept_from_activations(inputs, activations[self.split_point], concept_id, k)
+
+    @check_fitted
+    def top_k_inputs_for_concept_from_activations(
+        self,
+        inputs: ModelInput,
+        activations: LatentActivations,
         concepts_indices: int | list[int] | Literal["all"] = "all",
         k: int = 5,
     ) -> dict[int, list[tuple[str, float]]]:
         """
-        Retrieves the top-k most important tokens/words/clauses/phrases related to a given concept.
+        Retrieves the top-k most important input elements (tokens/words/clauses/phrases) related to a given concept.
+        This version of the method uses pre-computed activations
 
         Args:
-            inputs (List[List[str]]): TODO: clarify the supported types of inputs
-            activations (Float[torch.Tensor, "n l d"]): The activations to use for the analysis. They should correspond to the inputs. Shape: (n, l, d)
+            inputs (ModelInput): The input data, which can be a string, a list of tokens/words/clauses/sentences
+                or a dataset.
+            activations (LatentActivations): The activations to use for the analysis. They should correspond to the inputs. Shape: (n, l, d)
             concepts_indices (int | list[int] | Literal["all"]): The concept index (or list of concepts indices) to analyze. Defaults to "all".
             k (int): The number of important textual elements to retrieve. Defaults to 5.
 
@@ -215,104 +284,132 @@ class AbstractConceptExplainer(ABC):
 
         return interpretation_dict
 
+    @check_fitted
     def input_concept_attribution(
         self,
         inputs: ModelInput,
-        concept: int | list[int],
+        concept: int,
         attribution_method: type[AttributionExplainer],
         **attribution_kwargs,
     ) -> list[float]:
-        """
-        Computes the attribution of each input to a given concept.
+        """Attributes model inputs for a selected concept.
 
         Args:
-            inputs: The input data, which can be a string, a list of tokens/words/clauses/sentences, or a dataset.
-            concept: The concept index (or list of concepts indices) to analyze.
-            attribution_method: The method to use for attribution analysis.
+            inputs (ModelInput): The input data, which can be a string, a list of tokens/words/clauses/sentences
+                or a dataset.
+            concept (int): Index identifying the position of the concept of interest (score in the
+                `ConceptActivations` tensor) for which relevant input elements should be retrieved.
+            attribution_method: The attribution method to obtain importance scores for input elements.
 
         Returns:
             A list of attribution scores for each input.
         """
-        print(inputs, concept, attribution_method)
-        raise NotImplementedError()
+        raise NotImplementedError("Input-to-concept attribution method is not implemented yet.")
 
 
-class ConceptBottleneckExplainer(AbstractConceptExplainer):
-    """
-    The concept bottleneck explainer can encode activations into concepts and decode concepts into activations.
+class ConceptBottleneckExplainer(AbstractConceptExplainer[BaseDictionaryLearning], ABC):
+    """A concept bottleneck explainer wraps a `concept_model` that should be able to encode activations into concepts
+    and decode concepts into activations.
+
+    We use the term "concept bottleneck" loosely, as the latent space can be overcomplete compared to activation
+        space, as in the case of sparse autoencoders.
+
+    We assume that the concept model follows the structure of an [`overcomplete.BaseDictionaryLearning`](https://github.com/KempnerInstitute/overcomplete/blob/24568ba5736cbefca4b78a12246d92a1be04a1f4/overcomplete/base.py#L10)
+    model, which defines the `encode` and `decode` methods for encoding and decoding activations into concepts.
 
     Attributes:
-        splitted_model (ModelSplitterPlaceholder): The model to apply the explanation on. It should be splitted between at least two parts.
-        split (str): The split in the model where the concepts are encoded from.
-        concept_encoder_decoder (ConceptEncoderDecoder): Concept encoder-decoder
-        fitted (bool): Whether the model has been fitted
-        _differentiable_concept_encoder (bool): Whether the concept encoder is differentiable.
-        _differentiable_concept_decoder (bool): Whether the concept decoder is differentiable.
+        model_with_split_points (ModelWithSplitPoints): The model to apply the explanation on.
+            It should have at least one split point on which `concept_model` can be fitted.
+        split_point (str): The split point used to train the `concept_model`.
+        concept_model ([BaseDictionaryLearning](https://github.com/KempnerInstitute/overcomplete/blob/24568ba5736cbefca4b78a12246d92a1be04a1f4/overcomplete/base.py#L10)): The model used to extract concepts from the
+            activations of  `model_with_split_points`. The only assumption for classes inheriting from this class is
+            that the `concept_model` can encode activations into concepts with `encode_activations`.
+        is_fitted (bool): Whether the `concept_model` was fit on model activations.
+        has_differentiable_concept_encoder (bool): Whether the `encode_activations` operation is differentiable.
+        has_differentiable_concept_decoder (bool): Whether the `decode_concepts` operation is differentiable.
     """
 
-    _differentiable_concept_decoder: bool = False
-
-    def encode_activations(
-        self, activations: LatentActivations | dict[str, LatentActivations], **kwargs
-    ) -> ConceptsActivations:
-        """
-        Encode the given activations using the concept encoder-decoder.
+    def __init__(
+        self,
+        model_with_split_points: ModelWithSplitPoints,
+        concept_model: BaseDictionaryLearning,
+        split_point: str | None,
+    ):
+        """Initializes the concept explainer with a given splitted model.
 
         Args:
-            activations (LatentActivations | dict[str, LatentActivations]): The activations to encode.
+            model_with_split_points (ModelWithSplitPoints): The model to apply the explanation on.
+                It should have at least one split point on which a concept explainer can be trained.
+            concept_model ([BaseDictionaryLearning](https://github.com/KempnerInstitute/overcomplete/blob/24568ba5736cbefca4b78a12246d92a1be04a1f4/overcomplete/base.py#L10)): The model used to extract concepts from
+                the activations of `model_with_split_points`.
+            split_point (str | None): The split point used to train the `concept_model`. If None, tries to use the
+                split point of `model_with_split_points` if a single one is defined.
+        """
+        super().__init__(model_with_split_points, concept_model, split_point)
+        self.has_differentiable_concept_decoder = False
+
+    def __repr__(self):
+        return dedent(f"""\
+            {self.__class__.__name__}(
+                split_point={self.split_point},
+                concept_model={type(self.concept_model).__name__},
+                is_fitted={self.is_fitted},
+                has_differentiable_concept_encoder={self.has_differentiable_concept_encoder},
+                has_differentiable_concept_decoder={self.has_differentiable_concept_decoder},
+            )""")
+
+    @check_fitted
+    def encode_activations(self, activations: LatentActivations) -> torch.Tensor:  # ConceptActivations
+        """Encode the given activations using the `concept_model` encoder.
+
+        Args:
+            activations (LatentActivations): The activations to encode.
 
         Returns:
-            ConceptsActivations: The encoded activations.
+            The encoded concept activations.
         """
-        assert self.fitted, "Concept explainer has not been fitted yet."
+        self.verify_activations({self.split_point: activations})
+        return self.concept_model.encode(activations)  # type: ignore
 
-        inputs: LatentActivations
-        inputs, _ = self.verify_activations(activations)
-        inputs = inputs.to(self.concept_encoder_decoder.device)
-
-        return self.concept_encoder_decoder.encode(inputs, **kwargs)
-
-    def decode_concepts(self, concepts: ConceptsActivations) -> LatentActivations:
-        """
-        Decode the given concepts using the concept encoder-decoder.
+    @check_fitted
+    def decode_concepts(self, concepts: ConceptsActivations) -> torch.Tensor:  # LatentActivations
+        """Decode the given concepts using the `concept_model` decoder.
 
         Args:
             concepts (ConceptsActivations): The concepts to decode.
 
         Returns:
-            LatentActivations: The decoded activations.
+            The decoded model activations.
         """
-        assert self.fitted, "Concept explainer has not been fitted yet."
-        concepts = concepts.to(self.concept_encoder_decoder.device)
-        return self.concept_encoder_decoder.decode(concepts)
+        return self.concept_model.decode(concepts)  # type: ignore
 
+    @check_fitted
     def get_dictionary(self) -> torch.Tensor:
-        """
-        Get the dictionary learned by the concept encoder-decoder.
+        """Get the dictionary learned by the fitted `concept_model`.
 
         Returns:
-            torch.Tensor: The learned dictionary.
+            torch.Tensor: A `torch.Tensor` containing the learned dictionary.
         """
-        assert self.fitted, "Concept explainer has not been fitted yet."
-        return self.concept_encoder_decoder.get_dictionary()
+        return self.concept_model.get_dictionary()  # type: ignore
 
+    @check_fitted
     def concept_output_attribution(
-        self, concepts: ConceptsActivations, attribution_method: type[AttributionExplainer], **attribution_kwargs
-    ):
-        """
-        Computes the attribution of each concept to a given example.
+        self,
+        inputs: ModelInput,
+        concepts: ConceptsActivations,
+        target: int,
+        attribution_method: type[AttributionExplainer],
+        **attribution_kwargs,
+    ) -> list[float]:
+        """Computes the attribution of each concept for the logit of a target output element.
 
         Args:
-            concepts: The concepts to analyze.
-            attribution_method: The method to use for attribution analysis.
+            inputs (ModelInput): An input datapoint for the model.
+            concepts (torch.Tensor): Concept activation tensor.
+            target (int): The target class for which the concept output attribution should be computed.
+            attribution_method: The attribution method to obtain importance scores for input elements.
 
         Returns:
             A list of attribution scores for each concept.
-
-        Raises:
-            NotImplementedError: If the method is not implemented.
         """
-        assert self.fitted, "Concept explainer has not been fitted yet."
-        raise NotImplementedError(
-            f"Concept output attribution method {concepts}, {attribution_method} is not implemented yet."
-        )
+        raise NotImplementedError("Concept-to-output attribution method is not implemented yet.")
