@@ -32,123 +32,80 @@ import pytest
 import torch
 from overcomplete import optimization as oc_opt
 from overcomplete import sae as oc_sae
-from torch import nn
-from torch.nn import functional as F
+from pytest import fixture
+from transformers import AutoModelForMaskedLM
 
-from interpreto.commons.model_wrapping.model_splitter import ModelSplitterPlaceholder
-from interpreto.concepts.methods.overcomplete_cbe import (
+from interpreto.commons.model_wrapping.model_with_split_points import ModelWithSplitPoints
+from interpreto.concepts import (
     OvercompleteDictionaryLearning,
-    OvercompleteMethods,
+    OvercompleteOptimClasses,
     OvercompleteSAE,
+    OvercompleteSAEClasses,
 )
 
 DEVICE = "cuda" if torch.cuda.is_available() else "cpu"
 
-
-class EasilySplittableModel(nn.Module):
-    """
-    Dummy model with two parts for testing purposes
-    """
-
-    def __init__(self, input_size: int = 10, hidden_size: int = 20, output_size: int = 2):
-        super().__init__()
-        self.fc1 = nn.Linear(input_size, hidden_size)
-        self.fc2 = nn.Linear(hidden_size, hidden_size)
-        self.fc3 = nn.Linear(hidden_size, hidden_size)
-        self.fc4 = nn.Linear(hidden_size, output_size)
-
-    def input_to_latent(self, x: torch.Tensor) -> torch.Tensor:
-        x = F.relu(self.fc1(x))
-        return F.relu(self.fc2(x))
-
-    def end_model(self, x: torch.Tensor) -> torch.Tensor:
-        x = F.relu(self.fc3(x))
-        return F.relu(self.fc4(x))
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        x = self.input_to_latent(x)
-        return self.end_model(x)
+ALL_CONCEPT_METHODS = list(OvercompleteSAEClasses) + list(OvercompleteOptimClasses)
 
 
-def test_verify_activations():
-    """
-    Test verify_activations
-    """
-    n_samples = 20
-    input_size = 5
-    hidden_size = 10
-    n_concepts = 7
+@fixture
+def encoder_lm_splitter() -> ModelWithSplitPoints:
+    return ModelWithSplitPoints(
+        "huawei-noah/TinyBERT_General_4L_312D",
+        split_points=[],
+        model_autoclass=AutoModelForMaskedLM,  # type: ignore
+    )
 
-    model = EasilySplittableModel(input_size=input_size, hidden_size=hidden_size, output_size=2)
-    split = "input_to_latent"
-    splitted_model = ModelSplitterPlaceholder(model, split)
 
-    inputs = torch.randn(n_samples, input_size)
+def test_cbe_fit_failure_cases(encoder_lm_splitter: ModelWithSplitPoints):
+    """Test failure cases in matching the CBE with a ModelWithSplitPoints"""
+    encoder_lm_splitter.split_points = [
+        "cls.predictions.transform.LayerNorm",
+        "bert.encoder.layer.1",
+        "bert.encoder.layer.3.attention.self.query",
+    ]
 
-    # different splits between activations and model
-    activations = {
-        "input_to_latent": splitted_model.get_activations(inputs),
-        "end_model": inputs,
-    }
-    with pytest.raises(ValueError):
-        cbe = OvercompleteDictionaryLearning(splitted_model, oc_opt.NMF, n_concepts=n_concepts)
-        cbe.fit(activations, split="input_to_latent")
-
-    # several splits in activations and model but none specified
-    splitted_model.splits = ["input_to_latent", "end_model"]
-    # conflict with activations
-    with pytest.raises(ValueError):
-        cbe = OvercompleteDictionaryLearning(splitted_model, oc_opt.NMF, n_concepts=n_concepts)
-        cbe.fit(activations)
-    # conflict with activations
-    with pytest.raises(ValueError):
-        cbe = OvercompleteDictionaryLearning(splitted_model, oc_opt.NMF, n_concepts=n_concepts)
-        cbe.fit(activations[split])
+    # Raise when no split is provided and the model has more than one split
+    with pytest.raises(ValueError, match="If the model has more than one split point"):
+        cbe = OvercompleteDictionaryLearning(encoder_lm_splitter, oc_opt.NMF, nb_concepts=2)
+        assert not cbe.is_fitted
 
 
 @pytest.mark.slow
-def test_overcomplete_cbe():
-    """
-    Test OvercompleteSAE and OvercompleteDictionaryLearning
-    """
-    n_samples = 20
-    input_size = 5
-    hidden_size = 10
-    n_concepts = 7
+def test_overcomplete_cbe(encoder_lm_splitter: ModelWithSplitPoints):
+    """Test OvercompleteSAE and OvercompleteDictionaryLearning"""
 
-    model = EasilySplittableModel(input_size=input_size, hidden_size=hidden_size, output_size=2)
-    split = "input_to_latent"
-    splitted_model = ModelSplitterPlaceholder(model, split)
-
-    inputs = torch.randn(n_samples, input_size)
-    activations = splitted_model.get_activations(inputs)
-    assert activations[split].shape == (n_samples, hidden_size)
+    txt = ["Hello, my dog is cute", "The cat is on the [MASK]"]
+    split = "bert.encoder.layer.1"
+    nb_concepts = 3
+    encoder_lm_splitter.split_points = split
+    activations = encoder_lm_splitter.get_activations(txt, select_strategy="flatten")
+    assert activations[split].shape == (16, 312)
 
     # iterate over all methods from the namedtuple listing them
-    for method in OvercompleteMethods:
+    for method in ALL_CONCEPT_METHODS:
         if issubclass(method.value, oc_sae.SAE):
-            cbe = OvercompleteSAE(splitted_model, method.value, n_concepts=n_concepts, device=DEVICE)
-            cbe.fit(activations, nb_epochs=1, batch_size=n_samples // 2, device=DEVICE)
+            cbe = OvercompleteSAE(encoder_lm_splitter, method.value, nb_concepts=nb_concepts, device=DEVICE)
+            cbe.fit(activations, nb_epochs=1, batch_size=1, device=DEVICE)
         else:
             cbe = OvercompleteDictionaryLearning(
-                splitted_model,
+                encoder_lm_splitter,
                 method.value,
-                n_concepts=n_concepts,
+                nb_concepts=nb_concepts,
                 device=DEVICE,
             )
             cbe.fit(activations)
+        try:
+            assert hasattr(cbe, "concept_model")
+            assert hasattr(cbe, "model_with_split_points")
+            assert cbe.is_fitted
+            assert cbe.split_point == split
+            assert hasattr(cbe, "has_differentiable_concept_encoder")
+            assert hasattr(cbe, "has_differentiable_concept_decoder")
 
-        assert hasattr(cbe, "concept_encoder_decoder")
-        assert hasattr(cbe, "splitted_model")
-        assert cbe.fitted
-        assert cbe.split == split
-        assert hasattr(cbe, "_differentiable_concept_encoder")
-        assert hasattr(cbe, "_differentiable_concept_decoder")
-
-        concepts = cbe.encode_activations(activations)
-        assert concepts.shape == (n_samples, n_concepts)
-        reconstructed_activations = cbe.decode_concepts(concepts)
-        assert reconstructed_activations.shape == (n_samples, hidden_size)
-
-
-test_overcomplete_cbe()  # TODO: remove
+            concepts = cbe.encode_activations(activations[cbe.split_point])
+            assert concepts.shape == (16, nb_concepts)
+            reconstructed_activations = cbe.decode_concepts(concepts)
+            assert reconstructed_activations.shape == (16, 312)
+        except Exception as e:
+            raise AssertionError(f"Error with {method}") from e
